SIP PN: Prevent overlapping modifications of pn_ebr_filters template
The @pn_ebr_filters templates are actually allocated in shared memory and
can be modified by multiple SIP workers concurrently, leading to
occasional strange routing of INVITEs.  The fix is to simply allocate
the objects in private memory instead.

Credits to Ondrej Jirman <megi@xff.cz> for documenting, troubleshooting
and providing an initial fix!

Related to #3394