b2b_entitites: do not unref cancelled cell if not ours
If a CANCEL is caught by the b2b_entities but it is not part of a b2b
dialog, it should not unref the cell, otherwise after the script is
executed, it will be unreffed again, leading to a double unref.

Many thanks to Voxtronic for spotting this issue!