ratelimit: Avoid replicating "zero" pipes excessively
This patch makes it so any time a pipe's `.counter` value reaches 0, it
will only get broadcasted to neighbouring cluster nodes exactly 3 more
times, after which it will be completely skipped from being included in
the replication packet.  This counter is then reset to 3 as soon as the
pipe's value is incremented locally.