[clusterer] Fix error handling when failing to load nodes
On failing to add a node (to the cluster), continue, as time as that node is not the current node.
And if giving up on whole cluster info loading, be sure we clean up and properly report this to the above layer - the last thing we should do is to report success (on load) but have partial, unconsistent data.

Fixes #3473