clusterer: Prevent startup delays from causing aborted syncs
This patch makes it so the 'seed_fallback_interval' modparam
time-counting is done in a way that *ignores* any random startup delays
(e.g. mod init stuff which may take time such as loading drouting,
dispatcher, rtpproxy, etc.).

Previously, the effect of this issue was that some modules (e.g. usrloc)
would randomly *not* sync data on startup, depending if the startup
delay happened to be lesser or greater than the 'seed_fallback_interval'.