rtpengine: add set in the E_RTPENGINE_STATUS event
When having the same node in multiple sets, this event will be triggered
multiple times, but there is no way of figuring out for which set it has
been triggered. Adding the set allows you to do a better management of
the event, and also filter it better.

Many thanks to Wayne Davies from Five9 for reporting this issue!