dialog: accurate match of outgoing legs
The initial code assumes that for each transaction branch, we serially
create a dialog leg - therefore, when a reply is received for a specific
branch, we match the leg based on its index.
This logic breaks when having a branch that has no outgoing leg/message
(i.e. when push notification is used), because the dialog legs id gets
offsetted by the number of "blind" branches, thus we are no longer able
to properly match the reply with the leg.
This fix makes sure that when creating a let, it has the same index as
the transaction's branch id, hence the reply is properly matched.