db_sqlite: Remove infinite loops when a database lock occurs.
The proper way to handle a locked database is to use a busy timeout.
Previously, the behavior resulted in an infinite lock for the process
attempting to write to the locked database. The default value for the
busy timeout has been set to 500 ms to avoid immediate errors when short
database locks occur.