permissions: Use linked list for group storage
If an address/subnet is defined in multiple groups, `get_source_group`
may return a different group than it did previously.  Previously, it
returned the group with the lowest group ID where the match occurred
(due to how the data was stored).  Therefore, I am using a linked list
for groups instead of a hash table, so we do not have to look up every
group to find the smallest group ID.  This change is necessary for
backward compatibility.