[tm] t_lookup_ident() must not mess up with T
As t_lookup_ident() returns the found transaction via a parameter, it must not touch T. Everywhere where it is used, the code does not need/rely on T, but is always using only the trasaction from the returned param.
Setting the T is actually creating problem, as it it should set/restor T after a lookup.
This is an alternative fix to #3502

(cherry picked from commit e00f63e0f7a58cdd914bf54b9555c3d4e6a47486)