Added enhanced support for per msg branch variables
The msg branches are the branches created prior to t_relay(), with the set of destinations where the request should be parallel forked.
Available vars:
	$msg.branch.uri[br_idx]
	$msg.branch.duri[br_idx]
	$msg.branch.path[br_idx]
	$msg.branch.q[br_idx]
	$msg.branch.flags[br_idx]
	$msg.branch.flag(name)[br_idx]
	$msg.branch.socket[br_idx]
	$msg.branch.attr(name)[br_idx]
The indexing of the branches starts from 0 where 0 is the branch corresponding to RURI. The additional branches starts from idx 1.
What is new here is the concept of named attrs attached to each branch. These attrs can be used after the branch is created, all the way to the t_relay() moment.