dialog: avoid creating dialog multiple times if deleted
When an update command is received, the dialog is searched: before this
commit, if the dialog was not found, `dlg` was returned as `null`, and a
new dialog, **with the same coordinates** was created.
This commit makes sure that the dialog is no longer created if already
in deleted state, thus avoiding creating the dialog multiple times with
the same coordinates.

(cherry picked from commit 22b5a6b4158a92441aee33707f4b0ca8977a5a51)