SDP ops: Fix edge-case with missing ending LF
OpenSIPS no longer throws an error on SDP bodies with missing
line-termination sequence on the very last line.  According to SDP ABNF
specs, this is bad syntax, but we accept it and move on.

Moreover, assuming the SDP is dirty (at least one edit), the output SDP
($sdp) will always be normalized to include the ending LF sequence.