permissions: Small bugfix in DB schema
It seems bitmask values of 128 could not be stored in the table (e.g.
IPv6 addresses, not subnets), as TINYINT (char) can only support the
[-128, 127) interval.

As a fix, we silently bump the column to SMALLINT (2 bytes), solving the
issue but also not causing any breakage when pulling new packages
(e.g. no table version update, so we can backport this).