Fix memory leaks in stir_shaken module
The stir_shaken module would leak memory any time a signature was added
to a reply instead of the request or returned in a variable. This commit
makes sure the Identity header lump does not get duplicated and freed
only once. Also fixed possible memory leaks in case of issues while
adding a date header to a request without one.